refactor(xgoruntime): separate asset materialization
